requirement.txt
streamlit
pyspellchecker